DecInput and EncOutput -> uint8_t
Comm module should drop in at this point.

Added packing/unpacking functionality to encoder and decoder
Updated tests

Packing and Unpacking eats a lot of cycles, slows the Encoder and
Decoder down by about 50%. Right now it's implemented using the Pack()
and Unpack() calls. Hardcoding or careful casting could yield a big
speedup. The current implementation is not endianness-dependent,
however.

Had to modify MutexBuffer pop calls to take a 'multiple' argument.
It's very inconvenient if the Decoder Pop()s a number of entries
not divisible by the packed word size (5 bytes, in this case). If that
happened, the Decoder would have to keep track of the word fragment.
Instead of adding additional storage to the Decoder class for this
purpose, it was relatively simple to add a 'multiple' argument to the
Pop call itself. This argument is optional, so existing code is not
affected (multiple defaults to 1).

+++ bddriver/src/common/DriverTypes.h
@@ -103,7 +103,7 @@ struct DecOutput {
   unsigned int core_id;
   unsigned int time_epoch;
 };
-typedef uint64_t DecInput;
+typedef uint8_t DecInput;
 
 // encoder
 struct EncInput {
@@ -111,7 +111,7 @@ struct EncInput {
   unsigned int leaf_id;
   uint32_t payload;
 };
-typedef uint32_t EncOutput;
+typedef uint8_t EncOutput;
 
 } // bddriver
 } // pystorm
